Python/requirements.txt
flask==2.3.2
flask-cors==4.0.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
numpy==2.0.0
pandas==2.2.2
scikit-learn==1.4.2